Squashed commit of the following:
    net/mld: Checksum calculation needs to know the full size of the IPv6 header, including the size of the extension headers.  The payload size in the IPv64 header must include the extension headers (not considered part of the header in this case).  Fixes a few additional errors in size, endian-ness and checksum calculations.  Wireshark now reports the the outgoing Report has a good checksum.

    net/mld/mld_query.c:  Add a cast to assure that the left shift does not overflow.

+++ net/mld/mld_query.c
@@ -311,7 +311,7 @@ int mld_query(FAR struct net_driver_s *dev,
    * Query of length 26 octets) MUST be silently ignored.
    */
 
-  mldsize = ipv6->len[0] << 8 | ipv6->len[1];
+  mldsize = (unsigned int)ipv6->len[0] << 8 | ipv6->len[1];
   if (mldsize == sizeof(struct mld_mcast_listen_report_v1_s))
     {
       mldv1 = true;
